Fixes the organization of the App classes
This change is a step foward in the organization of this code repository. It renames the App classes so they are more discriptive, and also removes every hardcoded file path in the .java files. Lastly, it moves common code to an abstract app class, so the app classes are more DRY conformant.